feat(HW-885): Refactor transaction-address screen layout
